major modularity improvement for RulesAction, Rating and Rules modules
