+++ apps/processor/lib/processor/action.ex
@@ -1,4 +1,4 @@
-defmodule Processor.Action do
+defmodule Processor.RulesAction do
   @moduledoc """
   This is the `Action` module.
   This module receives action info and rules and this information generates an action structure.
@@ -8,53 +8,9 @@ defmodule Processor.Action do
   Transforms information based on action types to an action structure.
   """
   def to_action(info, rules) do
-    case info["action_type"] do
-      "pull_request" -> pull_request_to_action(info, rules["pull_request"])
-      "instagram_action" -> instagram_post_to_action(info, rules["instagram_action"])
-    end
-  end
-
-  defp instagram_post_to_action(info, rules) do
-    dharma = Processor.Rating.rate_instagram_post(info, rules)
-
-    %{
-      "type" => "action",
-      "action_type" => info["action_type"],
-      "title" => info["post"]["title"],
-      "user" => info["post"]["user"],
-      "stories" => info["post"]["stories"],
-      "dharma" => dharma
-    }
-  end
-
-  # Serialize a pull_request action type to an action structure.
-  defp pull_request_to_action(info, rules) do
-    dharma = Processor.Rating.rate_pull_request(info, rules)
-
-    %{
-      "type" => "action",
-      "action_type" => info["action_type"],
-      "owner" => info["owner"],
-      "repo" => info["repo"],
-      "title" => info["pull"]["title"],
-      "number_of_lines" => info["pull"]["additions"],
-      "user" => info["pull"]["user"]["login"],
-      "is_reviewed" => evaluate_reviews(info["reviews"]),
-      "commits" => info["pull"]["commits"],
-      "dharma" => dharma,
-      "closed_at" => info["pull"]["closed_at"],
-      "created_at" => info["pull"]["created_at"]
-    }
-  end
-
-  # Based on review value checks if it was reviewed or not.
-  defp evaluate_reviews(reviews) do
-    case reviews
-         |> Enum.filter(&(&1 != "COMMENTED"))
-         |> List.last() do
-      "APPROVED" -> "Reviewed"
-      "CHANGES_REQUESTED" -> "Changes Requested and not added"
-      _ -> "Unreviewed"
-    end
+    action_type = info["action_type"]
+    mod = String.to_atom("#{__MODULE__}.#{Macro.camelize(action_type)}")
+    fun = String.to_atom(action_type)
+    apply(mod, fun, [info, rules[action_type]])
   end
 end
