+++ apps/processor/test/processor/rules_action_test.exs
@@ -1,7 +1,7 @@
-defmodule ActionTest do
+defmodule RulesActionTest do
   use ExUnit.Case
   require Logger
-  doctest Processor.Action
+  doctest Processor.RulesAction
 
   test "Examplify processing an instagram action" do
     info = %{
@@ -16,12 +16,12 @@ defmodule ActionTest do
 
     rules = %{
       "instagram_action" => %{
-        "is_reviewed_instagram" => %{"is_reviewed" => 20, "not_reviewed" => 0},
+        "is_reviewed" => %{"is_reviewed" => 20, "not_reviewed" => 0},
         "number_of_stories" => %{"1" => 50, "2" => 100, "4" => 200}
       }
     }
 
-    action = Processor.Action.to_action(info, rules)
+    action = Processor.RulesAction.to_action(info, rules)
 
     Logger.info(inspect(action))
